+++ CHANGELOG.md
@@ -1,2 +1,18 @@
-- v2023-11-10.1800; initial version
-- v2023-11-11.1800; cleaned up code; github release
+```
+v1.0.0; 2025-08-27
+- stable v1.0.0 release
+- enforce Jotti's 250MB max file limit
+- added upload progress bar
+- added HTTP client timeout to avoid hangs
+- added non-zero exit on rate limit
+- tidied up logic in URL, filename, directory parsing
+```
+```
+v2023-11-11.1800
+- cleaned up code
+- github release
+```
+```
+v2023-11-10.1800
+- initial version
+```